Room database unit tests

+++ app/src/main/java/spr2024/cs2340/group9/studysync/database/Assignments.java
@@ -14,14 +14,22 @@ public class Assignments {
     private static AssignmentDao assignmentDao;
 
     /**
+     * Private empty constructor to force static methods.
+     */
+    private Assignments() {}
+
+    /**
      * Initialize object.
      * @param applicationContext context of the current application; use getApplicationContext() to get
      */
-    public Assignments(Context applicationContext) {
+    public static void init(Context applicationContext) {
         if (db == null) {
             db = Room.databaseBuilder(applicationContext, AssignmentDatabase.class,
-                    "Assignment").build();
+                    "Assignment")
+                    .allowMainThreadQueries()
+                    .build();
             assignmentDao = db.assignmentDao();
+            Assignment.currentId = assignmentDao.getId() + 1;
         }
     }
 
@@ -29,7 +37,10 @@ public class Assignments {
      * Get all assignments.
      * @return assignments
      */
-    public Assignment[] getAll() {
+    public static Assignment[] getAll() {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         return assignmentDao.getAll();
     }
 
@@ -37,7 +48,10 @@ public class Assignments {
      * Insert assignments into database.
      * @param assignments assignments
      */
-    public void insert(Assignment... assignments) {
+    public static void insert(Assignment... assignments) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         assignmentDao.insert(assignments);
     }
 
@@ -45,7 +59,10 @@ public class Assignments {
      * Delete an assignment from the database.
      * @param assignment assignment to delete
      */
-    public void delete(Assignment assignment) {
+    public static void delete(Assignment assignment) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         assignmentDao.delete(assignment);
     }
 
@@ -53,7 +70,10 @@ public class Assignments {
      * Delete an assignment from the database based on the id.
      * @param id id of assignment to delete
      */
-    public void delete(int id) {
+    public static void delete(int id) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         assignmentDao.delete(id);
     }
 
@@ -62,7 +82,10 @@ public class Assignments {
      * @param id assignment id
      * @return assignment
      */
-    public Assignment get(int id) {
+    public static Assignment get(int id) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         return assignmentDao.get(id);
     }
 
@@ -71,7 +94,10 @@ public class Assignments {
      * @param courseId course id
      * @return assignments
      */
-    public Assignment[] getCourse(int courseId) {
+    public static Assignment[] getCourse(int courseId) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         return assignmentDao.getCourse(courseId);
     }
 
@@ -81,20 +107,30 @@ public class Assignments {
      * @param endDate end of time frame
      * @return assignments
      */
-    public Assignment[] getBetween(Date startDate, Date endDate) {
-        return assignmentDao.getBetween(startDate.getTime(), endDate.getTime());
+    public static Assignment[] getBetween(long startDate, long endDate) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
+        return assignmentDao.getBetween(startDate, endDate);
     }
 
     /**
-     * Get all assignments with a specific ordering scheme.
-     * @param order ordering scheme
+     * Get all assignments with due dates within a time frame.
+     * @param startDate start of time frame
+     * @param endDate end of time frame
      * @return assignments
      */
-    public Assignment[] getOrder(Order order) throws IllegalArgumentException {
-        if (order == Order.START_TIME || order == Order.END_TIME) {
-            throw new IllegalArgumentException("Assignments cannot be ordered by START_TIME or" +
-                    "END_TIME.");
+    public static Assignment[] getBetween(Date startDate, Date endDate) {
+        return getBetween(startDate.getTime(), endDate.getTime());
+    }
+
+    /**
+     * Clear Assignment table.
+     */
+    public static void clear() {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
         }
-        return assignmentDao.getAll(order.columnName);
+        assignmentDao.clear();
     }
 }
