+++ app/src/main/java/spr2024/cs2340/group9/studysync/database/Courses.java
@@ -4,9 +4,6 @@ import android.content.Context;
 
 import androidx.room.Room;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-
 /**
  * Class to interface between the DAO and the rest of the code
  */
@@ -16,15 +13,23 @@ public class Courses {
     private static CourseTimeDao courseTimeDao;
 
     /**
+     * Private constructor to enforce static usage.
+     */
+    private Courses() {}
+
+    /**
      * Create a new Courses object.
      * @param applicationContext context of the current application; use getApplicationContext() to get
      */
-    public Courses(Context applicationContext) {
+    public static void init(Context applicationContext) {
         if (db == null) {
             db = Room.databaseBuilder(applicationContext, CourseDatabase.class, "Course")
+                    .allowMainThreadQueries()
                     .build();
             courseDao = db.courseDao();
             courseTimeDao = db.courseTimeDao();
+            Course.currentId = courseDao.getId() + 1;
+            CourseTime.currentId = courseTimeDao.getId() + 1;
         }
     }
 
@@ -32,10 +37,13 @@ public class Courses {
      * Get all courses.
      * @return courses
      */
-    public Course[] getAll() {
+    public static Course[] getAll() {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         Course[] courses = courseDao.getAll();
         for (Course course : courses) {
-            course.courseTimes = toTimePeriod(courseTimeDao.get(course.id));
+            course.setCourseTimes(toTimePeriod(courseTimeDao.get(course.id)));
         }
         return courses;
     }
@@ -44,21 +52,28 @@ public class Courses {
      * Insert courses.
      * @param courses courses
      */
-    public void insert(Course... courses) {
-        courseDao.insert(courses);
-        ArrayList<CourseTime> courseTimes = new ArrayList<>();
-        for (Course course: courses) {
-            courseTimes.addAll(Arrays.asList(toCourseTime(course.id, course.courseTimes)));
+    public static void insert(Course... courses) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
         }
-        CourseTime[] arr = new CourseTime[courseTimes.size()];
-        courseTimeDao.insert(courseTimes.toArray(arr));
+        for (Course course: courses) {
+            courseTimeDao.deleteCourse(course.id);
+            if (course.getCourseTimes() == null || course.getCourseTimes().length == 0) {
+                continue;
+            }
+            courseTimeDao.insert(toCourseTime(course.getCourseTimes()));
+        }
+        courseDao.insert(courses);
     }
 
     /**
      * Delete a course
      * @param course course
      */
-    public void delete(Course course) {
+    public static void delete(Course course) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         courseDao.delete(course);
         courseTimeDao.deleteCourse(course.id);
     }
@@ -68,9 +83,12 @@ public class Courses {
      * @param courseId course id
      * @return course
      */
-    public Course get(int courseId) {
+    public static Course get(int courseId) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         Course course = courseDao.get(courseId);
-        course.courseTimes = toTimePeriod(courseTimeDao.get(courseId));
+        course.setCourseTimes(toTimePeriod(courseTimeDao.get(courseId)));
         return course;
     }
 
@@ -79,10 +97,13 @@ public class Courses {
      * @param courseIds course ids to get courses
      * @return courses
      */
-    public Course[] get(int[] courseIds) {
+    public static Course[] get(int[] courseIds) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         Course[] courses = courseDao.get(courseIds);
         for (Course course: courses) {
-            course.courseTimes = toTimePeriod(courseTimeDao.get(course.id));
+            course.setCourseTimes(toTimePeriod(courseTimeDao.get(course.id)));
         }
         return courses;
     }
@@ -94,18 +115,31 @@ public class Courses {
      * @return courses in the time frame
      * @throws IllegalArgumentException if the startTime is after the endTime
      */
-    public Course[] getBetween(RecurringTime startTime, RecurringTime endTime) throws IllegalArgumentException {
-        if (startTime.compareTo(endTime) > 0) {
+    public static Course[] getBetween(int startTime, int endTime) throws IllegalArgumentException {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
+        if (startTime > endTime) {
             throw new IllegalArgumentException(String.format("The start time (%s) cannot be before " +
                     "the end time (%s).", startTime, endTime));
         }
-        int[] courseIds = courseTimeDao.getCourseIdBetween(startTime.getMinutesSinceSunday(),
-                endTime.getMinutesSinceSunday());
+        int[] courseIds = courseTimeDao.getCourseIdBetween(startTime, endTime);
         Course[] courses = courseDao.get(courseIds);
         for (Course course: courses) {
-            course.courseTimes = toTimePeriod(courseTimeDao.get(course.id));
+            course.setCourseTimes(toTimePeriod(courseTimeDao.get(course.id)));
         }
         return courses;
+    }
+
+    /**
+     * Get courses within a specific time frame.
+     * @param startTime beginning of the time frame, inclusive
+     * @param endTime end of the time frame, inclusive
+     * @return courses in the time frame
+     * @throws IllegalArgumentException if the startTime is after the endTime
+     */
+    public static Course[] getBetween(RecurringSlot startTime, RecurringSlot endTime) throws IllegalArgumentException {
+        return getBetween(startTime.getMinutesSinceSunday(), endTime.getMinutesSinceSunday());
     }
 
     /**
@@ -114,26 +148,34 @@ public class Courses {
      * @return courses on the specified day of the week
      * @throws IllegalArgumentException if the day of the week is invalid
      */
-    public Course[] getOnDay(int dayOfWeek) throws IllegalArgumentException {
+    public static Course[] getOnDay(int dayOfWeek) throws IllegalArgumentException {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
         if (dayOfWeek < 0 || dayOfWeek > 6) {
             throw new IllegalArgumentException(String.format("Day of week %d is invalid. Day of " +
                     "week must be in the range [0, 6], where 0 is Sunday and 6 is Saturday.",
                     dayOfWeek));
         }
-        return getBetween(new RecurringTime(dayOfWeek, 0, 0),
-                new RecurringTime(dayOfWeek, 23, 59));
+        return getBetween(new RecurringSlot(dayOfWeek, 0, 0),
+                new RecurringSlot(dayOfWeek, 23, 59));
     }
 
     /**
      * Private helper class to convert time periods to course times
-     * @param courseId course id for the time periods
-     * @param timePeriods time periods
+     * @param timeSlots time periods
      * @return course times
      */
-    private CourseTime[] toCourseTime(int courseId, TimePeriod[] timePeriods) {
-        CourseTime[] times = new CourseTime[timePeriods.length];
-        for (int i = 0; i < timePeriods.length; i++) {
-            times[i] = timePeriods[i].getCourseTime(courseId);
+    private static CourseTime[] toCourseTime(TimeSlot[] timeSlots) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
+        if (timeSlots == null || timeSlots.length == 0) {
+            return null;
+        }
+        CourseTime[] times = new CourseTime[timeSlots.length];
+        for (int i = 0; i < timeSlots.length; i++) {
+            times[i] = timeSlots[i].toCourseTime();
         }
         return times;
     }
@@ -143,11 +185,27 @@ public class Courses {
      * @param courseTimes course times
      * @return time periods
      */
-    private TimePeriod[] toTimePeriod(CourseTime[] courseTimes) {
-        TimePeriod[] periods = new TimePeriod[courseTimes.length];
+    private static TimeSlot[] toTimePeriod(CourseTime[] courseTimes) {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
+        if (courseTimes == null || courseTimes.length == 0) {
+            return null;
+        }
+        TimeSlot[] periods = new TimeSlot[courseTimes.length];
         for (int i = 0; i < courseTimes.length; i++) {
-            periods[i] = new TimePeriod(courseTimes[i]);
+            periods[i] = new TimeSlot(courseTimes[i]);
         }
         return periods;
+    }
+
+    /**
+     * Clears table.
+     */
+    public static void clear() {
+        if (db == null) {
+            throw new IllegalStateException("Database function can not be run because the database has not been initialized.");
+        }
+        courseDao.clear();
     }
 }
